Small changes to the satp handling functions

+++ src/memory/virtmem.c
@@ -37,7 +37,7 @@ Error initKernelVirtualMemory() {
     return simpleError(SUCCESS);
 }
 
-void setVirtualMemory(int asid, PageTable* page_table, bool fence) {
+void setVirtualMemory(uint16_t asid, PageTable* page_table, bool fence) {
     if (page_table == NULL) {
         setSatpCsr(0);
     } else {
@@ -48,7 +48,7 @@ void setVirtualMemory(int asid, PageTable* page_table, bool fence) {
     }
 }
 
-uint64_t satpForMemory(int asid, PageTable* page_table) {
+uint64_t satpForMemory(uint16_t asid, PageTable* page_table) {
     return 8L << 60 | (uint64_t)asid << 44L | (uintptr_t)page_table >> 12;
 }
