+++ src/memory/virtmem.h
@@ -15,8 +15,8 @@ void setSatpCsr(uint64_t satp);
 
 void addressTranslationFence(int asid);
 
-void setVirtualMemory(int asid, PageTable* page_table, bool fence);
+void setVirtualMemory(uint16_t asid, PageTable* page_table, bool fence);
 
-uint64_t satpForMemory(int asid, PageTable* page_table);
+uint64_t satpForMemory(uint16_t asid, PageTable* page_table);
 
 #endif
